Defect Number: #81479
Description Of Defect: For Item Type CR, Points Earned is 0, even though CR scoring is not done in score details accordion of view status page.

How The Issue Was Solved: Acceptance criteria of related story is updated and the new enhancement requested is implemented as documentation change.

Hints To QA For Testing: Login to TAS for TASC Readiness user and check for response and score details in 'Score Details' accordion of View Status page for individual rosters.

Developer Comments:  The TASC Readiness customer should have the "View_Response_Result" configuration and default value is 'T'. This feature should be checked in various TAS UI supported browsers.