~Defect/Story Number: OAS-3322 ~Description Of Defect/Story: LLO RP - Perform Validation with BMT during Test Session Delete ~Who Did Code Review: Souvik Das ~Code Files Changes: /dep-base-ctrl/src/com/ctb/control/db/TestRoster.java[Branch : OAS94] /dep-base-ctrl/src/com/ctb/util/OperationStatus.java[Branch : OAS94] /dep-testadmin-ctrl/src/com/ctb/control/testAdmin/ScheduleTest.java[Branch : OAS94] /dep-testadmin-ctrl/src/com/ctb/control/testAdmin/ScheduleTestImpl.java[Branch : OAS94] /web-testsessioninfo/src/testsessionApplicationResource.properties[Branch : OAS94] /web-testsessioninfo/src/sessionOperation/SessionOperationController.java[Branch : OAS94] /web-testsessioninfo/WebContent/resources/js/sessionList.js[Branch : OAS94] /web-testsessioninfo/WebContent/sessionOperation/oas_message.jsp[Branch : OAS94]
~Impacted Areas/Regression Areas: application-testsessioninfo. Delete session operation in Manage Session page.

~How The Issue Was Solved: Configuration "LLO_RP_Customer" has been used and based on that configuration the BMT Validation webservice is called.Upon validation from this webservice it is decided that whether a session can be deleted from OAS.

~Hints To QA For Testing: 	Give the new configuration to a customer with default value 'T' and then Schedule, Edit, Copy Sessions.Delete that session  and check as per the BMTValidation WebService response.

~Developer Comments:  a) The customer must have the configuration "LLO_RP_Customer" with default value 'T'.
b) If validation fails for any one of the students present in the session, then the session will not be deleted.
c) The customer must have entry in the customer_resource table with "RESOURCE_TYPE_CODE" column value as "BMTAPIURL" and "RESOURCE_URI" column value as "https://oastest.ctb.com/dummyrest/getCompletionStatus"
d) For QA testing purpose a dummy BMT webservice has been written and it uses the below mentioned logic. If the EXT_PIN1 of a student is null or alphanumeric then the BMTvalidation will fail, i.e. the student cannot be deleted. If EXT_PIN1 is numeric then BMTvalidation will succeed, i.e. the student can be deleted.