Defect/Story Number: 	OAS-1456 Description Of Defect/Story: 	TASC TEI Scoring MSS: Implement error handling and retry logic for TEI Scoring via MSS Who Did Code Review: 	Somenath Chakroborti Code Files Changes: 	/dep-scoring-ctrl/src/com/ctb/lexington/domain/score/ResponseReplayer.java [Branch : OAS94] 	/dep-scoring-ctrl/src/com/ctb/lexington/domain/score/scorer/calculator/MosaicResponseProcessCalculator.java [Branch : OAS94] 	/dep-scoring-ctrl/src/com/ctb/lexington/util/MosaicAuthentication.java [Branch : OAS94] 	/dep-scoring-ctrl/src/com/ctb/lexington/util/MosaicConstantUtils.java [Branch : OAS94]
	also config.properties file updated in all the IRS QA clusters
Impacted Areas/Regression Areas: 	Scoring module for TASC Operational and TASC Readiness.
How The Issue Was Solved: 	If any error occurs at the time of MSS WS call then the roster data is stored in a table named "SCORING_INVOKE_LOG". There is a scheduler which fetches this data and retries the scoring of those rosters.
Hints To QA For Testing: 	Schedule a TASC Operational and TASC Readiness test. Complete exam in TDC. Check scoring in the IRS Database. Check the "SCORING_INVOKE_LOG" table for an erroneous roster and check the retry process for the roster is working.
Developer Comments:  	a) This is a progressive work for Retry Count Issue.
	b) The OAS-DAS item mapping data has not been received yet for all the TE items. So, the scoring will be done for only the currently loaded items. For other items the item score value will be null.
	c) A new column "FIELD_TEST" has been added to TASC_ITEM_FACT table for TE FT item identification.